+++ src/__sdk.js
@@ -1,1 +1,1 @@
-module.exports = '4.7.3';
+module.exports = '4.10.0';