+++ test/e2e/code/world.js
@@ -1,4 +1,8 @@
-const { defineSupportCode } = require('cucumber');
+const { setWorldConstructor } = require('cucumber');
+const { getUrl } = require('./helpers/https');
+const { waitForEmail } = require('./helpers/maildrop');
+const { getRandomUsername } = require('./helpers/random-data');
+const testConfig = require('./test-config');
 
 /**
  * Cucumber creates an instance of this class for each scenario.
@@ -26,11 +30,11 @@ class PeerioAppWorld {
         });
     };
 
-    login = async () => {
+    login = async (username, passphrase) => {
         await this.libs.prombservable.asPromise(this.ice.socket, 'connected', true);
         const u = new this.ice.User();
-        u.username = this.username;
-        u.passphrase = this.passphrase;
+        u.username = username || this.username;
+        u.passphrase = passphrase || this.passphrase;
         this.ice.User.current = u;
         await u.login();
         return this.waitForAccountDataInit();
@@ -44,11 +48,54 @@ class PeerioAppWorld {
         await asPromiseNegative(this.ice.User.current, 'quota', null);
         console.log('Account init: waiting for settings to load');
         await asPromise(this.ice.User.current.settings, 'loaded', true);
+        console.log('Account init: waiting for tofuStore to load');
+        await asPromise(this.ice.tofuStore, 'loaded', true);
+        console.log('Account init: waiting for contactStore to load');
+        await asPromise(this.ice.contactStore.myContacts.loaded, true);
         console.log('Account init: waiting self contact info to load');
         await this.ice.contactStore.currentUser.ensureLoaded();
     }
+
+    confirmPrimaryEmail = async (emailAddress) => {
+        const email = await waitForEmail(emailAddress, testConfig.primaryEmailConfirmSubject);
+        const url = testConfig.emailConfirmUrlRegex.exec(email.body)[1];
+        await getUrl(url);
+    }
+
+    createAccount = async (username, email, isTestAccount = false) => {
+        await this.libs.prombservable.asPromise(this.ice.socket, 'connected', true);
+
+        const u = new this.ice.User();
+        u.username = username || getRandomUsername();
+        u.email = email || `${u.username}@${testConfig.emailDomain}`;
+        u.firstName = 'Firstname';
+        u.lastName = 'Lastname';
+        u.locale = 'en';
+        u.passphrase = testConfig.defaultPassphrase;
+        this.ice.User.current = u;
+        if (!isTestAccount) {
+            this.username = u.username;
+            this.passphrase = u.passphrase;
+        } else {
+            this.testAccount = {
+                username: u.username,
+                passphrase: u.passphrase,
+                email: u.email
+            };
+        }
+
+        console.log(`creating user username: ${this.username} passphrase: ${this.passphrase}`);
+
+        await u.createAccountAndLogin();
+        console.log('Account created, waiting for initialization.');
+        await this.waitForAccountDataInit();
+    };
+
+    createTestAccount = async (username = null, email = null) => {
+        return this.createAccount(username, email, true);
+    }
 }
 
-defineSupportCode(({ setWorldConstructor }) => {
-    setWorldConstructor(PeerioAppWorld);
-});
+
+setWorldConstructor(PeerioAppWorld);
+