+++ test/e2e/run.sh
@@ -3,14 +3,15 @@ mkdir -p ./test-results/e2e
 
 if [ $CI ]
 then
-    tags='not @wip and not @long'
+    tags='not @wip and not @long and not @off'
 else
-    tags='not @wip'
+    tags='not @wip and not @off'
 fi
-node --expose-gc ./node_modules/.bin/cucumber.js test/e2e/spec \
+node --expose-gc ./node_modules/.bin/cucumber-js test/e2e/spec \
         -r test/e2e/code \
-        --compiler js:babel-register \
+        --require-module babel-register \
         --format node_modules/cucumber-pretty \
         --format usage:./test-results/e2e/usage.txt \
         --format json:./test-results/e2e/result.json \
-        --tags "$tags"
+        --tags "$tags" \
+        --exit